+++ src/Authenticator/BearerAuthenticator.php
@@ -18,17 +18,17 @@ use Symfony\Component\Security\Http\Authenticator\Passport\Badge\UserBadge;
 use Symfony\Component\Security\Http\Authenticator\Passport\Passport;
 use Symfony\Component\Security\Http\Authenticator\Passport\SelfValidatingPassport;
 
+/**
+ * @internal
+ */
 class BearerAuthenticator extends AbstractAuthenticator implements LoggerAwareInterface
 {
     use LoggerAwareTrait;
 
-    private BearerUserProviderInterface $userProvider;
-    private OIDCUserSessionProviderInterface $userSession;
-
-    public function __construct(OIDCUserSessionProviderInterface $userSession, BearerUserProviderInterface $userProvider)
+    public function __construct(
+        private readonly OIDCUserSessionProviderInterface $userSessionProvider,
+        private readonly BearerUserProviderInterface $userProvider)
     {
-        $this->userProvider = $userProvider;
-        $this->userSession = $userSession;
     }
 
     public function supports(Request $request): ?bool
@@ -60,7 +60,7 @@ class BearerAuthenticator extends AbstractAuthenticator implements LoggerAwareIn
         $passport = new SelfValidatingPassport(new UserBadge($user->getUserIdentifier(), function ($token) use ($user) {
             return $user;
         }));
-        $passport->setAttribute('relay_user_session_provider', $this->userSession);
+        $passport->setAttribute('relay_user_session_provider', $this->userSessionProvider);
 
         return $passport;
     }