+++ src/OIDCProvider/OIDError.php
@@ -4,6 +4,9 @@ declare(strict_types=1);
 
 namespace Dbp\Relay\CoreConnectorOidcBundle\OIDCProvider;
 
+/**
+ * @internal
+ */
 class OIDError extends \Exception
 {
 }